us/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
attrs==24.2.0
blinker==1.8.2
cachelib==0.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
crawlerdetect==0.1.7
Deprecated==1.2.15
dicttoxml==1.7.16
dnspython==2.7.0
emoji==2.14.0
et-xmlfile==1.1.0
filelock==3.16.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-Limiter==3.9.2
frozenlist==1.4.1
geoip2==4.8.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
limits==3.14.1
markdown-it-py==3.0.0
MarkupSafe==3.0.1
maxminddb==2.6.2
mdurl==0.1.2
mongomock==4.2.0.post1
multidict==6.1.0
openpyxl==3.1.5
ordered-set==4.1.0
packaging==24.1
pluggy==1.5.0
propcache==0.2.0
pycountry==24.6.1
Pygments==2.18.0
pymongo==4.11.1
pytest==8.3.3
pytest-flask==1.3.0
pytest-mock==3.14.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-file==2.1.0
requests-mock==1.12.1
rich==13.9.4
ruff==0.9.6
sentinels==1.0.0
setuptools==75.1.0
tldextract==5.1.2
typing_extensions==4.12.2
ua-parser==0.18.0
urllib3==2.2.3
user-agents==2.2.0
validators==0.34.0
Werkzeug==3.0.6
wrapt==1.17.0
yarl==1.15.2
qrcode==7.4.2
Pillow==10.3.0
